Add player scoring functionality and update game events

+++ src/main/scala/de/htwg/se/thirtyone/model/Player.scala
@@ -2,6 +2,6 @@ package de.htwg.se.thirtyone.model
 
 case class Player(
     hasKnocked: Boolean = false, 
-    points: Int = 0,
+    points: Double = 0,
     playersHealth: Int = 3
 )
